+++ src/sampler_utils.py
@@ -55,7 +55,7 @@ def _sample_from_solid_probability(
     size = len(probability.shape)
     subkeys = split(subkey, size)
     idx = size - 1
-    samples = jnp.zeros((size,), dtype=jnp.int32)
+    samples = jnp.zeros((size,), dtype=jnp.int8)
     for _ in range(size):
         probability = probability.reshape((-1, 16))
         marginal_prob = probability.sum(0)
